--- Athlete.java
@@ -1,16 +1,0 @@
-import greenfoot.*;  // (World, Actor, GreenfootImage, Greenfoot and MouseInfo)
-
-/**
- * Write a description of interface StudentAthlete here.
- * 
- * @author Shivani Sridhar 
- * @version August 29, 2023
- * Shivani, Mohith, Samika, Eric
- */
-public interface Athlete 
-{
-    // method signatures - implement the signature below in your own class. Make sure to
-    //                     match the parameter list and return type
-    
-    public void mySport(String mySentence, int numberOfSports);
-}
