+++ Student.java
@@ -69,100 +69,5 @@ public abstract class Student extends Actor
         mySeatX=getX();
         mySeatY=getY();
     }
-
-    
- /**
- * Description of epicDanceMove()
- * The epicDanceMove method incorporates elements from both disco() and doABarrelRoll() methods
- * to create a unique dance move. The student object will zigzag, flash colors, and spin.
- *
- * @author Mohit, Eric, Shivani, Samika
- * @version 1.2z Build 9002 Aug 30, 2023.
- */
-public void epicDanceMove() {
-    for (int i = 0; i <= 5; i++) {
-        int originalWidth = getImage().getWidth();
-        int originalHeight = getImage().getHeight();
-        
-        // Start by doing the zigzag
-        setLocation(this.getX() + 5, this.getY() + 5);
-        Greenfoot.delay(3);
-        setLocation(this.getX() - 5, this.getY() - 5);
-        Greenfoot.delay(3);
-        
-        // Turn 30 degrees
-        turn(30);
-        Greenfoot.delay(5);
-
-        // Change the image to a random color (assuming you have a method to do this)
-        changeRandomColor();
-        Greenfoot.delay(5);
-
-        // Resize the image to 1.5x its original dimensions
-        getImage().scale((int)(originalWidth * 1.5), (int)(originalHeight * 1.5));
-        Greenfoot.delay(5);
-        
-        // Turn -30 degrees
-        turn(-30);
-        Greenfoot.delay(5);
-
-        // Restore the image to its original dimensions and color
-        getImage().scale(originalWidth, originalHeight);
-        resetColor();
-        Greenfoot.delay(5);
-    }
-}
-    /** Description of disco()
- * The disco method makes the student do a dance where they shake, increase and decrease size, and change transparency
- * 
- * @author Aditya D
- * @author Raman A
- * @author Conrad A
- * @author Vishnu D
- * @version 1.0z Build 9000 Aug 30, 2023.
- */
-    public void disco() {
-        
-        for (int i=0; i <=15; i++){
-            int originalWidth = getImage().getWidth();
-            int originalHeight = getImage().getHeight();
-            setRotation(15);
-            getImage().setTransparency(255);
-            Greenfoot.delay(10);
-            getImage().scale(originalWidth*2,originalWidth*2);
-            Greenfoot.delay(10);
-            getImage().scale(originalWidth,originalHeight);
-            Greenfoot.delay(10);
-            setRotation(-15);
-            Greenfoot.delay(10);
-            getImage().scale(originalWidth*2,originalWidth*2);
-            Greenfoot.delay(10);
-            getImage().scale(originalWidth,originalHeight);
-            Greenfoot.delay(10);
-            getImage().setTransparency(0);
-            Greenfoot.delay(10);
-        }
-    }
-    /*
-     * doABarrelRoll() will cause a student object to do a barrel roll forwards and backwards. 
-     * This can be used during the animation when a student is clicked on.  
-     * Created by Tharun Selvakumar, Sid Mathur, Aidan Spies, and Himanshu M.
-     */
-    public void doABarrelRoll(){
-        for(int i = 0; i < 8; i++){
-            setLocation(this.getX() + 1, this.getY());
-            //Greenfoot.delay(1);
-            turn(45);
-            Greenfoot.delay(7);
-        }
-        for(int i = 0; i < 8; i++){
-            setLocation(this.getX() - 1, this.getY()); 
-            //Greenfoot.delay(1);
-            turn(-45);
-            Greenfoot.delay(7);
-        }
-
-    }
  
-}
 }
