+++ StudentDeskGroup.java
@@ -19,9 +19,8 @@ public class StudentDeskGroup extends Actor
      * the 'Act' or 'Run' button gets pressed in the environment.
      */
     public void act() 
-    {
+    {   
+        
         // Add your action code here.
     }    
-    
-    
 }
